Update categoria.js
No se pueden borrar categorías con libros asociados.

+++ src/api/categoria.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 const CategoriaModel = require('../models/categoria')
+const LibroModel = require("../models/libro");
 
 router.post('/', async(req, res, next) => {
     const categoria = new CategoriaModel({
@@ -40,16 +41,25 @@ router.get('/:id', async(req, res, next) => {
 });
 
 
-router.delete('/:id', async(req, res, next) => {
-    const { id } = req.params;
-    try {
-        const categoriaBorrada = await CategoriaModel.findByIdAndDelete(id);
-        res.status(200).json(categoriaBorrada);
-        console.log('Se borro correctamente');
-    } catch (error) {
-        res.status(413).send({ mensaje: 'Error inesperado, Categoria con libros asociados no se puede eliminar, No existe Categoria indicada' });
-        next(error);
+router.delete("/:id", async (req, res, next) => {
+  const { id } = req.params;
+  try {
+    const categoria = await CategoriaModel.findById(id);
+    const libro = await LibroModel.find({ categoria_id: categoria.id });
+    if (libro == "") {
+      const categoriaBorrada = await CategoriaModel.findByIdAndDelete(id);
+      res.status(200).json(categoriaBorrada);
     }
+    res.status(413).send({
+      mensaje: "La categoria no se puede borrar ya que tiene libros asociados.",
+    });
+  } catch (error) {
+    res.status(413).send({
+      mensaje:
+        "Error inesperado, Categoria con libros asociados no se puede eliminar, No existe Categoria indicada",
+    });
+    next(error);
+  }
 });
 
 module.exports = router;
